change layout of show more button
